Stop mock hosts deploy failing so often
It's common for mock host deploy to do nothing, and so be very fast
(sub-minute). Then it tries to tag it with the same tag as the previous
deploy, which fails.

Adding seconds to the tag will stop this.

(cherry picked from commit be9eae283cf6e9a83eaefe3665f087402c3e2330)