Less wasteful dev redeploy strategy
Only force redeploys if the service is already on the desired tag

Previously each service was force redeploying whenever any branch built,
and in the case of mock host redeploying all services whenever any
branch built.

Now only redeploys the specific service on successful build of the
branch currently deployed for that service in dev.