Allow mock hosts to have static outbound ips
By putting them on a private subnet and disabling `AssignPublicIp` we
should force them to use the NAT gateway on the public subnets which has
a static elastic ip.

Only works for mock hosts with a network load balancer in front of them,
otherwise this will make them inaccessible from the internet. So only
mock hosts with static inbound IPs via an NLB can be in private subnets.